requirements.txt
fastapi
python-dotenv
supabase
psycopg[binary,pool]
pydantic
autumn
httpx
requests

